feature: add tqdm to follow progress. Refactored file strcture and changed .env
