+++ .env.example
@@ -5,7 +5,7 @@ YOLO_MODEL=src/yolo/darknet/cfg/yolov3.cfg
 YOLO_CLASSES=src/yolo/darknet/data/coco.names
 
 VIDEO=src/yolo/Cycling.mp4
-TRACKED_VIDEO=src/yolo/tracked_video.avi
+TRACKED_VIDEO_FOLDER=src/yolo/
 
 GITY_PATH=src/yolo/darknet/
