+++ .gitignore
@@ -162,3 +162,4 @@ cython_debug/
 #.idea/
 
 *.part
+tracked_video*.avi
